Develop (#2)
* Attempting to publish changes

* Fixed empty stack bug

Legacy code would duplicate a card if a stack was empty in order to fill it even if the empty stack was not the destination.

* Separated gameMode buttons into their own class

* game runs from same frame instead of creating a new frame

* Sound and Visual Update

Cards are now images. Added music and sound effects. Added Options menu on main menu that allows user to adjust sound of music or sound effects. Favorites now save and load automatically.

* Refactored code

moved UI out of the FlowerBed class and adjusted control and dependencies on StartMenu, UI, and GameMode

* continued refactoring

Moved mouse events to ui controller and set up solitaire to work with the new UI elements

* Fixed card drawing lag

legacy code in FlowerBedCardStack was lagging the cards.

* updated stat tracking

added total games, wins, last score, last time

* Bug fix and options menu improvement

Fixed bug with games auto ending after a move to the foundation. Added UI placement, card choice, and background choice options to the options menu.

* Fixed alerts

* Cards now load in with the selected backs

Co-authored-by: jacksonGiles <81192950+jacksonGiles@users.noreply.github.com>